Only autodetect chips if there is a unique chip found
The autodetect code was not verifying that only a single chip with
the given manufacturer and part ID exists in the registry.

This led to problems with some STM32 chips, as the part-number is
not unique for them. It might also be a problem for other chips.

This commit fixes this by returning an error if more than a single
chip is found, to ensure no chips are wrongly detected.