Use sticky overrun behaviour for improved JLink performance
To avoid protocol errors in case of WAIT or FAULT responses,
we use the *sticky overun* behaviour of the debugport.

This behaviour means that WAIT and FAULT responses are also followed
by the normal data phases, instead of stopping immediately like normal.

We cannot stop the data transmission on the JLink, so this helps
avoiding protocol errors and should improve performance.