Only clear HW breakpoints when necessary
Currently, we always try to clear all HW breakpoints when the
Session is dropped. Especially when there are communication problems,
this leads to a lot of noise in the log output.

With this commit, we only try to clear the breakpoints which were
actually set by probe-rs itself, when the session is dropped.

When attaching, we still clear all HW breakpoints to ensure the target
is in a known state.