riscv: try to use abstract commands for CSR access
Instead of using program buffer for accessing (almost) all CSR,
try to use abstract commands directly access CSR.
If a given CSR fails with abstract cmd err 2 (unsupported), the
program buffer is used as a fallback strategy.

The given read / write failure is also stored in a cache, and the
next try on the same register will use the program buffer directly.

This leads to a massive (almost 4x) speedup of e.g. discovering
breakpoints in get_available_breakpoint_units().