Remove AsMut<Probe> and AsRef<Probe>, fix some clippy lints
This was only used to get access to the target_reset_deassert function,
so it's simpler to just make that function available. It's also cleaner,
because access to the other Probe function is not necessary.

Also fix some clippy lints, and remove some lint changes which didn't
have any effect.