Fix 8 bit reads on ST-Link
Before this commit, reads of a single byte on ST-Link were extended
to a read of 2 bytes, to work around some issues with the ST-Link.

This is problematic when trying to read the last byte in
a memory region, the extended read will then read from
another part of memory, which can fail.

Reading just a single byte seems to work, but the buffer for returned
data has to be at least two bytes in size, otherwise a USB overflow
occurs.