DapAccess: remove port number from PortType::AccessPort.
PortType is used for DapAccess, which is the low-level trait. Users of
DapAccess must do the AP selection themselves (by writing to APSEL), so
impls of DapAccess don't need the AP number, they just need need to know
the port type so they can set the APnDP bit correctly.

Indeed, in the jlink and cmsis-dap impls the AP number was unused.