debugger: parse wire protocol correctly, only configure when requested
In the vscode extension, the name 'wire_protocol' is used, but the
struct expected 'protocol'.

Also, the debugger always used SWD, even if no protocol was specified.
Some probes only support JTAG; so this leads to problems.
Just not specifying a protocol is a better solution, then the probe
itself can use its default protocol.