Add Mutex back to GDB run method
In the last round of refactoring Mutex was removed from the run() function signature due to no longer needing within the GDB logic.  However, consumers of this API, such as cargo-embed, depend on being able to run multiple threads so the GDB code cannot take exclusive ownership of the Session object.

Signed-off-by: Ryan Fairfax <ryan@thefaxman.net>