Reduce NVM ranges for all MIMXRT10xx targets
If the NVM range is larger than the flash algorithm's range, the flash
algorithm filter (loader.rs:445 as of this commit) will not select that
algorithm. It's a curious behavior, because the NVM range seems large
enough to accomodate the flashing algorithm's needs.

The MIMXRT10xx targets have NVM regions that a larger than their flash
algorith ranges, so we can't flash MIMXRT10xx targets without changing
either the target definitions, or the filter. We're not ready to change
the filter.

Before this commit, the end of flash represented the maximum size of the
MIMXRT10xx FlexSPI regions. This commit reduces the range to match the
range of the associated flashing algorithm. By reducing the range, the
flash algorithm filter now selects a flashing algorithm.

We're not sure if supporting NVM region > flash algorithm ranges is
necessary. But with this commit in place, we can answer that question
later. If probe-rs ends up supporting the behavior, we could revert this
commit to make the entire flash range available.