Add `DebugProbeError::ReAttachRequired`
Devices such as the ATSAMD/E5x require that the probe re-attaches
after a Chip-Erase in case the device got unlocked in the process
because the new "unlocked" status only takes effect after the next
re-attach.

Add a new error `DebugProbeError::ReAttachRequired` that can be used
to signal this condition. It can either be handled directly by
re-attaching or just bubble it up the user who will re-try
the operation.