Improve logging in probe-rs-cli-util (#1464)
The logging was done in the format call, which is not the correct place.
Also, some logging calls where missing when the output was redirect to a
file, since indicatif does not print the messages in that case.

This PR also changes the log output target, the log messages are now
sent to stderr, not stdout. This is the default behaviour of the logger.
The log target could be changed back to stdout if desired.